--- src/common/LeftDrawer/index.jsx
@@ -1,53 +1,0 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import { withStyles } from '@material-ui/core/styles'
-import Drawer from '@material-ui/core/Drawer'
-import List from '@material-ui/core/List'
-import ListItem from '@material-ui/core/ListItem'
-import ListItemIcon from '@material-ui/core/ListItemIcon'
-import ListItemText from '@material-ui/core/ListItemText'
-import DashboardIcon from '@material-ui/icons/Dashboard'
-import Divider from '@material-ui/core/Divider'
-import { withRouter } from 'react-router-dom'
-import Accounts from '../../core/Accounts'
-import LinkTo from '../LinkTo'
-
-const styles = theme => ({
-  toolbar: theme.mixins.toolbar,
-  drawerPaper: {
-    width: 200
-  }
-})
-
-
-const LeftDrawer = ({ classes, match }) => (
-  <Drawer
-    elevation={3}
-    variant="permanent"
-    classes={{
-      paper: classes.drawerPaper
-    }}
-  >
-    <div className={classes.toolbar} />
-    <List>
-      <ListItem
-        button
-        key="Dashboard"
-        component={LinkTo('/dashboard')}
-        selected={match.path === '/dashboard'}
-      >
-        <ListItemIcon><DashboardIcon /></ListItemIcon>
-        <ListItemText primary="Dashboard" />
-      </ListItem>
-    </List>
-    <Divider />
-    <Accounts accountId={('accountId' in match.params) ? match.params.accountId : null} />
-  </Drawer>
-)
-
-LeftDrawer.propTypes = {
-  classes: PropTypes.object.isRequired,
-  match: PropTypes.object.isRequired
-}
-
-export default withStyles(styles)(withRouter(LeftDrawer))
