+++ src/store/user/actions.js
@@ -35,7 +35,10 @@ export const loadUserData = () => (dispatch, getState) => {
       pictureUrl: person.avatarUrl()
     }))
     storage.loadState().then((state) => {
-      dispatch(loadSettings((state || {}).settings))
+      dispatch(loadSettings({
+        ...(state || {}).settings,
+        overlayMessage: ('settings' in getState() ? getState().settings.overlayMessage : '')
+      }))
       dispatch(loadAccounts((state || {}).accounts))
       dispatch(loadTransactions((state || {}).transactions))
       dispatch(loadExchangeRates((state || {}).exchangeRates))
@@ -84,11 +87,17 @@ export const handleBlockstackLogin = () => (dispatch) => {
 }
 
 export const resetState = () => (dispatch, getState) => {
-  dispatch(loadSettings({
-    ...settingsInitialState,
-    currency: getState().settings.currency,
-    locale: getState().settings.locale
-  }))
+  const { settings } = getState().settings
+  if (settings === undefined) {
+    dispatch(loadSettings(settingsInitialState))
+  } else {
+    // Don't remove the existing currency and locale
+    dispatch(loadSettings({
+      ...settingsInitialState,
+      currency: settings.currency,
+      locale: settings.locale
+    }))
+  }
   dispatch(loadAccounts(accountsInitialState))
   dispatch(loadTransactions(transactionsInitialState))
   dispatch(loadExchangeRates(exchangeRatesInitialState))
